VoiceAssistant-master/requirements.txt
pyttsx3>=2.90
SpeechRecognition>=3.8.1
google-cloud-speech>=2.11.0
paho-mqtt>=1.6.1

portion>=2.2.0
mpmath>=1.2.1